Perbaikan Policies

+++ app/Http/Controllers/CustomerController.php
@@ -84,6 +84,8 @@ class CustomerController extends Controller
         {
         //
         // dd($request);
+        $this->authorize('update', $customer);
+
         $validateData = $request->validate([
             'nama_customer' => 'required|max:50',
             'alamat' => 'required|max:200',
@@ -118,6 +120,7 @@ class CustomerController extends Controller
     public function destroy(customer $customer)
     {
          //
+        $this->authorize('delete', $customer);
         $customer->delete();
         Session::flash('success','Data berhasil dihapus');
         return redirect()->back();
